requirements.txt
google-api-python-client==1.8.4
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
apscheduler==3.0
requests==2.23.0
numpy==1.18.4
pandas==1.0.3
pytz==2020.1
